[undertow] Extract logic from the conjure exception handling HttpHandler
This is a refactor, no logic or public API change.

Exception serialization logic moved from
`ConjureExceptionHandler.handleRequset(HttpServerExchange)` into
an internal utility function `ConjureExceptions.handle`.

This is a pre-req for async request processing, see #361
